Ensures files are added to watch before calling in callback (RT#115380)
Files that were modified in the callback were sometimes not added to the
watch list first (specifically in KQueue and Inotify2). This patch
ensures that files/directories are added to the watch before the
callback is called.

If you modified _process_events (a private method) this could break your
code.

Upgrade is recommended if you use the KQueue or Inotify2 backends.